block coder: ensure all dimensions and loop variables are 16 bit

+++ src/lib/core/t1/part1/block_coder/BlockCoder.h
@@ -68,7 +68,7 @@ private:
   /**
    * @brief cached block stride
    */
-  uint8_t stride_;
+  uint16_t stride_;
   /**
    * @brief cached block height
    */
